+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to
 - ⬆️(dependencies) update pydanticai and related dependencies
 - ⚙️(back) get carbon data from albert api provider
 - ✨(tools): update descriptions
+- ✨(auth) add silent OIDC login
 
 ### Fixed